Process Packets from Streams
handle_read handles the logic for piecing together packets.
start_packet_read and the future packets that will be reading from the
buffer are designed to fail grafefully if a full packet is not
available. This will be useful for bad connections and maybe UDP.

A simple Handshake test is in start_packet_read to test reading
different types from the buffer.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>