Packet Macro with Reading/Writing
This turned out better than I thought it would.
You can now create a packet and you only need to specify the arguments.
It'll throw compile time errors if you try to use a type that isn't
supported. Not all the type traits are implemented yet, and I still
need to do protocols.

Packets aren't bound to an id so that a packet struct can be shared
across protocols if it hasn't changed. That means small updates like
Java Edition Snapshots can reuse mose of the packets from the older
version.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>