Integrate Protocol into Connections
wow it worked first try

The packets are parsed using the current Protocol the Connection is
using. A Connection's protocol will always start on the most generic
version and then specialize. For Java Edition, it'll be the lowest
protocol we have until the client sends theirs. For Bedrock Edition,
we need to start on the RakNet protocol until the client sends their
protocol.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>